Initial attempt at mongo support
Some notes:

The port from SQL to mongo was as simple and literal as possible.
We can probably leverage some features of mongo to make it more
native to a sql-less DB.

Also, not all tests are passing yet. I'm not sure what approach I
want to take with performable_method. Separating it into two classes
seems overkill, but I don't want to overload that class with too many
specific constants.